+++ agentid/__init__.py
@@ -1,3 +1,3 @@
 """AgentAction package."""
 
-__version__ = "0.22.0rc2"
+__version__ = "0.22.0rc3"